+++ packages/runtime-common/index.ts
@@ -153,7 +153,7 @@ import type { CodeRef } from './code-ref';
 export type { CodeRef };
 
 export * from './code-ref';
-export * from './serializer';
+export * from './serializers';
 
 export type { CardDocument, SingleCardDocument } from './document-types';
 export type {
