+++ packages/runtime-common/serializers/image-size.ts
